pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sommelier"
version = "0.1.0"
description = "AI-powered Netflix content recommender — a conversational sommelier for streaming titles"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.47",
    "pandas>=2.1",
    "scikit-learn>=1.4",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
sommelier = "sommelier.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sommelier"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
